Move some of TestInstanceOf back to legend-pure
