+++ legend-engine-core/legend-engine-core-pure/legend-engine-pure-code-functions-unclassified/legend-engine-pure-runtime-java-extension-interpreted-functions-unclassified/src/test/java/org/finos/legend/pure/runtime/java/interpreted/function/base/meta/TestInstanceOf.java
@@ -14,10 +14,9 @@
 
 package org.finos.legend.pure.runtime.java.interpreted.function.base.meta;
 
+import org.finos.legend.engine.pure.code.core.functions.unclassified.base.meta.AbstractTestInstanceOf;
 import org.finos.legend.pure.m3.execution.FunctionExecution;
-import org.finos.legend.pure.m3.tests.function.base.meta.AbstractTestInstanceOf;
 import org.finos.legend.pure.runtime.java.interpreted.FunctionExecutionInterpreted;
-import org.junit.After;
 import org.junit.BeforeClass;
 
 public class TestInstanceOf extends AbstractTestInstanceOf
@@ -26,12 +25,6 @@ public class TestInstanceOf extends AbstractTestInstanceOf
     public static void setUp()
     {
         setUpRuntime(getFunctionExecution());
-    }
-
-    @After
-    public void cleanRuntime()
-    {
-        runtime.delete("fromString.pure");
     }
 
     protected static FunctionExecution getFunctionExecution()
